+++ lib/decorators/validate.decorator.ts
@@ -1,8 +1,9 @@
 import { applyDecorators, SetMetadata } from '@nestjs/common';
 import { RMQ_ROUTES_VALIDATE } from '../constants';
 
-export const Validate = (): MethodDecorator => {
-	return applyDecorators(
-		SetMetadata(RMQ_ROUTES_VALIDATE, true),
-	);
+export const RMQValidate = (): MethodDecorator => {
+	return applyDecorators(SetMetadata(RMQ_ROUTES_VALIDATE, true));
 };
+
+/** @depreceated */
+export const Validate = RMQValidate;