+++ lib/helpers/logger.ts
@@ -21,7 +21,7 @@ export class RQMColorLogger implements LoggerService {
 		const action = context.split(',')[0];
 		const topic = context.split(',')[1];
 		Logger.log(`${blueBright(action)} [${yellow(topic)}] ${white(msg)}`);
-		console.warn(`${blueBright(action)} [${yellow(topic)}] ${white(msg)}`)
+		console.warn(`${blueBright(action)} [${yellow(topic)}] ${white(msg)}`);
 	}
 	warn(message: any, context?: string): any {
 		Logger.warn(message, context);