+++ lib/rmq.explorer.ts
@@ -10,16 +10,15 @@ import { ExtendedMessage } from './classes/rmq-extended-message.class';
 import { RMQError } from './classes/rmq-error.class';
 import { IRouteOptions } from './interfaces/queue-meta.interface';
 import { validate } from 'class-validator';
+import { plainToClass } from 'class-transformer';
 
 @Injectable()
 export class RMQExplorer implements OnModuleInit {
-
 	constructor(
 		private readonly discoveryService: DiscoveryService,
 		private readonly metadataAccessor: RMQMetadataAccessor,
-		private readonly metadataScanner: MetadataScanner,
-	) {
-	}
+		private readonly metadataScanner: MetadataScanner
+	) {}
 
 	async onModuleInit() {
 		this.explore();
@@ -35,10 +34,8 @@ export class RMQExplorer implements OnModuleInit {
 			if (!instance || !Object.getPrototypeOf(instance)) {
 				return;
 			}
-			this.metadataScanner.scanFromPrototype(
-				instance,
-				Object.getPrototypeOf(instance),
-				(key: string) => this.lookupRMQRoute(instance, key),
+			this.metadataScanner.scanFromPrototype(instance, Object.getPrototypeOf(instance), (key: string) =>
+				this.lookupRMQRoute(instance, key)
 			);
 		});
 	}
@@ -54,26 +51,28 @@ export class RMQExplorer implements OnModuleInit {
 		this.attachEmitter(path, options, instance, methodRef);
 	}
 
-	private attachEmitter(path: string, options: IRouteOptions, instance: Record<string, Function>, methodRef: Function) {
+	private attachEmitter(
+		path: string,
+		options: IRouteOptions,
+		instance: Record<string, Function>,
+		methodRef: Function
+	) {
 		requestEmitter.on(path, async (msg: Message) => {
-			const messageParams: number[] =
-				this.metadataAccessor.getRMQMessageIndexes(Object.getPrototypeOf(instance), methodRef.name);
+			const messageParams: number[] = this.metadataAccessor.getRMQMessageIndexes(
+				Object.getPrototypeOf(instance),
+				methodRef.name
+			);
 			try {
-				const funcArgs = options?.msgFactory
-					? options.msgFactory(msg)
-					: RMQMessageFactory(msg);
+				let funcArgs = options?.msgFactory ? options.msgFactory(msg) : RMQMessageFactory(msg);
 				if (messageParams.length > 0) {
 					for (const param of messageParams) {
 						funcArgs[param] = new ExtendedMessage(msg);
 					}
 				}
+				funcArgs = this.transformRequest(instance, methodRef, funcArgs);
 				const error = await this.validateRequest(instance, methodRef, funcArgs);
 				if (error) {
-					responseEmitter.emit(
-						ResponseEmitterResult.error,
-						msg,
-						new RMQError(error, ERROR_TYPE.RMQ),
-					);
+					responseEmitter.emit(ResponseEmitterResult.error, msg, new RMQError(error, ERROR_TYPE.RMQ));
 					responseEmitter.emit(ResponseEmitterResult.ack, msg);
 					return;
 				}
@@ -84,7 +83,7 @@ export class RMQExplorer implements OnModuleInit {
 					responseEmitter.emit(
 						ResponseEmitterResult.error,
 						msg,
-						new RMQError(ERROR_UNDEFINED_FROM_RPC, ERROR_TYPE.RMQ),
+						new RMQError(ERROR_UNDEFINED_FROM_RPC, ERROR_TYPE.RMQ)
 					);
 				}
 			} catch (err) {
@@ -98,9 +97,21 @@ export class RMQExplorer implements OnModuleInit {
 		});
 	}
 
+	private transformRequest(instance: Record<string, Function>, methodRef: Function, funcArgs: any[]): any[] {
+		const transformMsg = this.metadataAccessor.getRMQTransformation(methodRef);
+		if (!transformMsg) {
+			return funcArgs;
+		}
+		const types = Reflect.getMetadata('design:paramtypes', Object.getPrototypeOf(instance), methodRef.name);
+		funcArgs[0] = plainToClass(types[0], funcArgs[0]);
+		return funcArgs;
+	}
+
 	private async validateRequest(
-		instance: Record<string, Function>, methodRef: Function, funcArgs: any[]
-		): Promise<string | undefined> {
+		instance: Record<string, Function>,
+		methodRef: Function,
+		funcArgs: any[]
+	): Promise<string | undefined> {
 		const validateMsg = this.metadataAccessor.getRMQValidation(methodRef);
 		if (!validateMsg) {
 			return;